More clean up of Albany 64bit Tpetra capabilities.
Disable Epetra in discretization classes in 64bit mode.
Add cee-compute011 to performance tests.

+++ src/Albany_DataTypes.hpp
@@ -60,7 +60,7 @@ typedef Sacado::Fad::DFad<MPType> MPFadType;
 //Tpetra typedefs
 typedef double                                      ST;
 #ifdef ALBANY_64BIT_INT
-typedef long long                                         GO;
+typedef long                                        GO;
 #else
 typedef int                                         GO;
 #endif
